requirements.txt
asttokens==2.2.1
attrs==22.2.0
backcall==0.2.0
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.2
confection==0.0.4
contourpy==1.0.7
cycler==0.11.0
cymem==2.0.7
dataclasses==0.6
debugpy==1.6.6
decorator==5.1.1
docopt==0.6.2
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl
executing==1.2.0
Flask==2.2.3
flatbuffers==23.3.3
focusedme==0.1.32
fonttools==4.39.0
humanfriendly==10.0
idna==3.4
imbalanced-learn==0.10.1
imblearn==0.0
ipykernel==6.21.2
ipython==8.10.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jupyter_client==8.0.3
jupyter_core==5.2.0
kiwisolver==1.4.4
langcodes==3.3.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mpmath==1.3.0
murmurhash==1.0.9
nest-asyncio==1.5.6
nltk==3.8.1
numpy==1.24.2
onnxruntime==1.14.1
packaging==23.0
pandas==1.5.3
parso==0.8.3
pathy==0.10.1
pickleshare==0.7.5
Pillow==9.4.0
platformdirs==3.0.0
playsound==1.3.0
preshed==3.0.8
prompt-toolkit==3.0.37
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
pydantic==1.10.5
Pygments==2.14.0
pyparsing==3.0.9
pyreadline3==3.4.1
python-dateutil==2.8.2
pytz==2022.7.1
pywin32==305
pyzmq==25.0.0
regex==2022.10.31
requests==2.28.2
scikit-learn==1.2.1
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
spacy==3.5.0
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
sympy==1.11.1
textblob==0.17.1
thinc==8.1.8
threadpoolctl==3.1.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.14
vaderSentiment==3.3.2
wasabi==1.1.1
wcwidth==0.2.6
Werkzeug==2.2.3
